pagination fully finsished except on team page

+++ AMTProject01/src/main/java/Services/Player/PlayerManager.java
@@ -28,7 +28,7 @@ public interface PlayerManager {
     public void Add(Player p);
     public List<Player> getPlayersFrom(Team t) ;
     
-    public List<Player> getPlayersFromMatchTeam1(int match_id) ;
-    public List<Player> getPlayersFromMatchTeam2(int match_id) ;
+    public List<Player> getPlayersFromMatchTeam1(long match_id) ;
+    public List<Player> getPlayersFromMatchTeam2(long match_id) ;
     
 }
